# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Benefits Enrollment MVP API
  version: 0.2.0
  description: Multi-tenant benefits enrollment API with role-based access controls.
servers:
  - url: http://localhost:4000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
    Role:
      type: string
      enum: [FULL_ADMIN, COMPANY_ADMIN, EMPLOYEE]
...
    AuthUser:
      type: object
      required: [id, email, role, tenantId]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/Role'
        tenantId:
          type: [string, 'null']
          format: uuid
        sessionId:
          type: [string, 'null']
          format: uuid
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/AuthUser'
    SignupInviteRequest:
      type: object
      required: [inviteCode, email, password]
      properties:
        inviteCode:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    PasswordResetRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    PasswordResetConfirmRequest:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
        newPassword:
          type: string
          minLength: 8
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        resetToken:
          type: string
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                error: Invalid or expired token
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid:
              value:
                error: Validation failed
                details:
                  - path: email
                    message: Invalid email
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                login:
                  value:
                    accessToken: jwt-token
                    refreshToken: refresh-token
                    user:
                      id: 8f530abe-7332-4df1-9727-a8527dbffb1e
                      email: platform-admin@example.com
                      role: FULL_ADMIN
                      tenantId: null
                      sessionId: 1f530abe-7332-4df1-9727-a8527dbffb1e
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/signup-invite:
    post:
      tags: [Auth]
      summary: Create account using invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupInviteRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token and issue new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke refresh token session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Logged out
  /auth/logout-all:
    post:
      tags: [Auth]
      summary: Revoke all sessions for current user
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logged out from all sessions
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/password-reset/request:
    post:
      tags: [Auth]
      summary: Request password reset token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          description: Password reset request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /auth/password-reset/confirm:
    post:
      tags: [Auth]
      summary: Confirm password reset using token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
      responses:
        '200':
          description: Password reset complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
...